+++ api/tasks.go
@@ -274,144 +274,6 @@ func (s *Spread) Canonicalize() {
 	}
 }
 
-// CheckRestart describes if and when a task should be restarted based on
-// failing health checks.
-type CheckRestart struct {
-	Limit          int            `mapstructure:"limit"`
-	Grace          *time.Duration `mapstructure:"grace"`
-	IgnoreWarnings bool           `mapstructure:"ignore_warnings"`
-}
-
-// Canonicalize CheckRestart fields if not nil.
-func (c *CheckRestart) Canonicalize() {
-	if c == nil {
-		return
-	}
-
-	if c.Grace == nil {
-		c.Grace = timeToPtr(1 * time.Second)
-	}
-}
-
-// Copy returns a copy of CheckRestart or nil if unset.
-func (c *CheckRestart) Copy() *CheckRestart {
-	if c == nil {
-		return nil
-	}
-
-	nc := new(CheckRestart)
-	nc.Limit = c.Limit
-	if c.Grace != nil {
-		g := *c.Grace
-		nc.Grace = &g
-	}
-	nc.IgnoreWarnings = c.IgnoreWarnings
-	return nc
-}
-
-// Merge values from other CheckRestart over default values on this
-// CheckRestart and return merged copy.
-func (c *CheckRestart) Merge(o *CheckRestart) *CheckRestart {
-	if c == nil {
-		// Just return other
-		return o
-	}
-
-	nc := c.Copy()
-
-	if o == nil {
-		// Nothing to merge
-		return nc
-	}
-
-	if o.Limit > 0 {
-		nc.Limit = o.Limit
-	}
-
-	if o.Grace != nil {
-		nc.Grace = o.Grace
-	}
-
-	if o.IgnoreWarnings {
-		nc.IgnoreWarnings = o.IgnoreWarnings
-	}
-
-	return nc
-}
-
-// The ServiceCheck data model represents the consul health check that
-// Nomad registers for a Task
-type ServiceCheck struct {
-	Id            string
-	Name          string
-	Type          string
-	Command       string
-	Args          []string
-	Path          string
-	Protocol      string
-	PortLabel     string `mapstructure:"port"`
-	AddressMode   string `mapstructure:"address_mode"`
-	Interval      time.Duration
-	Timeout       time.Duration
-	InitialStatus string `mapstructure:"initial_status"`
-	TLSSkipVerify bool   `mapstructure:"tls_skip_verify"`
-	Header        map[string][]string
-	Method        string
-	CheckRestart  *CheckRestart `mapstructure:"check_restart"`
-	GRPCService   string        `mapstructure:"grpc_service"`
-	GRPCUseTLS    bool          `mapstructure:"grpc_use_tls"`
-}
-
-// The Service model represents a Consul service definition
-type Service struct {
-	Id           string
-	Name         string
-	Tags         []string
-	CanaryTags   []string `mapstructure:"canary_tags"`
-	PortLabel    string   `mapstructure:"port"`
-	AddressMode  string   `mapstructure:"address_mode"`
-	Checks       []ServiceCheck
-	CheckRestart *CheckRestart `mapstructure:"check_restart"`
-	Connect      *ConsulConnect
-}
-
-func (s *Service) Canonicalize(t *Task, tg *TaskGroup, job *Job) {
-	if s.Name == "" {
-		s.Name = fmt.Sprintf("%s-%s-%s", *job.Name, *tg.Name, t.Name)
-	}
-
-	// Default to AddressModeAuto
-	if s.AddressMode == "" {
-		s.AddressMode = "auto"
-	}
-
-	// Canonicalize CheckRestart on Checks and merge Service.CheckRestart
-	// into each check.
-	for i, check := range s.Checks {
-		s.Checks[i].CheckRestart = s.CheckRestart.Merge(check.CheckRestart)
-		s.Checks[i].CheckRestart.Canonicalize()
-	}
-}
-
-type ConsulConnect struct {
-	SidecarService *ConsulSidecarService `mapstructure:"sidecar_service"`
-}
-
-type ConsulSidecarService struct {
-	Port  string
-	Proxy *ConsulProxy
-}
-
-type ConsulProxy struct {
-	Upstreams []*ConsulUpstream
-}
-
-type ConsulUpstream struct {
-	//FIXME Pointers?
-	DestinationName string `mapstructure:"destination_name"`
-	LocalBindPort   int    `mapstructure:"local_bind_port"`
-}
-
 // EphemeralDisk is an ephemeral disk object
 type EphemeralDisk struct {
 	Sticky  *bool
@@ -628,6 +490,9 @@ func (g *TaskGroup) Canonicalize(job *Job) {
 	}
 	for _, n := range g.Networks {
 		n.Canonicalize()
+	}
+	for _, s := range g.Services {
+		s.Canonicalize(nil, g, job)
 	}
 }
