+++ jobspec/parse.go
@@ -947,9 +947,12 @@ func parseTasks(jobName string, taskGroupName string, result *[]*api.Task, list 
 		}
 
 		if o := listVal.Filter("service"); len(o.Items) > 0 {
-			if err := parseServices(jobName, taskGroupName, &t, o); err != nil {
+			services, err := parseServices(jobName, taskGroupName, o)
+			if err != nil {
 				return multierror.Prefix(err, fmt.Sprintf("'%s',", n))
 			}
+
+			t.Services = services
 		}
 
 		// If we have config, then parse that
@@ -1287,8 +1290,8 @@ func parseGroupServices(jobName string, taskGroupName string, g *api.TaskGroup, 
 	return nil
 }
 
-func parseServices(jobName string, taskGroupName string, task *api.Task, serviceObjs *ast.ObjectList) error {
-	task.Services = make([]*api.Service, len(serviceObjs.Items))
+func parseServices(jobName string, taskGroupName string, serviceObjs *ast.ObjectList) ([]*api.Service, error) {
+	services := make([]*api.Service, len(serviceObjs.Items))
 	for idx, o := range serviceObjs.Items {
 		// Check for invalid keys
 		valid := []string{
@@ -1299,22 +1302,24 @@ func parseServices(jobName string, taskGroupName string, task *api.Task, service
 			"check",
 			"address_mode",
 			"check_restart",
+			"connect",
 		}
 		if err := helper.CheckHCLKeys(o.Val, valid); err != nil {
-			return multierror.Prefix(err, fmt.Sprintf("service (%d) ->", idx))
+			return nil, multierror.Prefix(err, fmt.Sprintf("service (%d) ->", idx))
 		}
 
 		var service api.Service
 		var m map[string]interface{}
 		if err := hcl.DecodeObject(&m, o.Val); err != nil {
-			return err
+			return nil, err
 		}
 
 		delete(m, "check")
 		delete(m, "check_restart")
+		delete(m, "connect")
 
 		if err := mapstructure.WeakDecode(m, &service); err != nil {
-			return err
+			return nil, err
 		}
 
 		// Filter checks
@@ -1322,33 +1327,236 @@ func parseServices(jobName string, taskGroupName string, task *api.Task, service
 		if ot, ok := o.Val.(*ast.ObjectType); ok {
 			checkList = ot.List
 		} else {
-			return fmt.Errorf("service '%s': should be an object", service.Name)
+			return nil, fmt.Errorf("service '%s': should be an object", service.Name)
 		}
 
 		if co := checkList.Filter("check"); len(co.Items) > 0 {
 			if err := parseChecks(&service, co); err != nil {
-				return multierror.Prefix(err, fmt.Sprintf("service: '%s',", service.Name))
+				return nil, multierror.Prefix(err, fmt.Sprintf("service: '%s',", service.Name))
 			}
 		}
 
 		// Filter check_restart
 		if cro := checkList.Filter("check_restart"); len(cro.Items) > 0 {
 			if len(cro.Items) > 1 {
-				return fmt.Errorf("check_restart '%s': cannot have more than 1 check_restart", service.Name)
+				return nil, fmt.Errorf("check_restart '%s': cannot have more than 1 check_restart", service.Name)
 			}
 			if cr, err := parseCheckRestart(cro.Items[0]); err != nil {
-				return multierror.Prefix(err, fmt.Sprintf("service: '%s',", service.Name))
+				return nil, multierror.Prefix(err, fmt.Sprintf("service: '%s',", service.Name))
 			} else {
 				service.CheckRestart = cr
 			}
 		}
 
-		task.Services[idx] = &service
+		// Filter connect
+		if co := checkList.Filter("connect"); len(co.Items) > 0 {
+			if len(co.Items) > 1 {
+				return nil, fmt.Errorf("connect '%s': cannot have more than 1 connect stanza", service.Name)
+			}
+
+			c, err := parseConnect(co.Items[0])
+			if err != nil {
+				return nil, multierror.Prefix(err, fmt.Sprintf("service: '%s',", service.Name))
+			}
+
+			service.Connect = c
+		}
+
+		services[idx] = &service
 	}
 
-	return nil
+	return services, nil
 }
 
+func parseConnect(co *ast.ObjectItem) (*api.ConsulConnect, error) {
+	valid := []string{
+		"native",
+		"sidecar_service",
+	}
+
+	if err := helper.CheckHCLKeys(co.Val, valid); err != nil {
+		return nil, multierror.Prefix(err, "connect ->")
+	}
+
+	var connect api.ConsulConnect
+	var m map[string]interface{}
+	if err := hcl.DecodeObject(&m, co.Val); err != nil {
+		return nil, err
+	}
+
+	delete(m, "sidecar_service")
+
+	if err := mapstructure.WeakDecode(m, &connect); err != nil {
+		return nil, err
+	}
+
+	var connectList *ast.ObjectList
+	if ot, ok := co.Val.(*ast.ObjectType); ok {
+		connectList = ot.List
+	} else {
+		return nil, fmt.Errorf("connect should be an object")
+	}
+
+	// Parse the sidecar_service
+	o := connectList.Filter("sidecar_service")
+	if len(o.Items) == 0 {
+		return &connect, nil
+	}
+	if len(o.Items) > 1 {
+		return nil, fmt.Errorf("only one 'sidecar_service' block allowed per task")
+	}
+
+	r, err := parseSidecarService(o.Items[0])
+	if err != nil {
+		return nil, fmt.Errorf("sidecar_service, %v", err)
+	}
+	connect.SidecarService = r
+
+	return &connect, nil
+}
+
+func parseSidecarService(o *ast.ObjectItem) (*api.ConsulSidecarService, error) {
+	valid := []string{
+		"port",
+		"proxy",
+	}
+
+	if err := helper.CheckHCLKeys(o.Val, valid); err != nil {
+		return nil, multierror.Prefix(err, "sidecar_service ->")
+	}
+
+	var sidecar api.ConsulSidecarService
+	var m map[string]interface{}
+	if err := hcl.DecodeObject(&m, o.Val); err != nil {
+		return nil, err
+	}
+
+	delete(m, "proxy")
+
+	dec, err := mapstructure.NewDecoder(&mapstructure.DecoderConfig{
+		DecodeHook:       mapstructure.StringToTimeDurationHookFunc(),
+		WeaklyTypedInput: true,
+		Result:           &sidecar,
+	})
+	if err != nil {
+		return nil, err
+	}
+	if err := dec.Decode(m); err != nil {
+		return nil, fmt.Errorf("foo: %v", err)
+	}
+
+	var proxyList *ast.ObjectList
+	if ot, ok := o.Val.(*ast.ObjectType); ok {
+		proxyList = ot.List
+	} else {
+		return nil, fmt.Errorf("sidecar_service: should be an object")
+	}
+
+	// Parse the proxy
+	po := proxyList.Filter("proxy")
+	if len(po.Items) == 0 {
+		return &sidecar, nil
+	}
+	if len(po.Items) > 1 {
+		return nil, fmt.Errorf("only one 'proxy' block allowed per task")
+	}
+
+	r, err := parseProxy(po.Items[0])
+	if err != nil {
+		return nil, fmt.Errorf("proxy, %v", err)
+	}
+	sidecar.Proxy = r
+
+	return &sidecar, nil
+}
+
+func parseProxy(o *ast.ObjectItem) (*api.ConsulProxy, error) {
+	valid := []string{
+		"upstreams",
+		"config",
+	}
+
+	if err := helper.CheckHCLKeys(o.Val, valid); err != nil {
+		return nil, multierror.Prefix(err, "proxy ->")
+	}
+
+	var proxy api.ConsulProxy
+
+	var listVal *ast.ObjectList
+	if ot, ok := o.Val.(*ast.ObjectType); ok {
+		listVal = ot.List
+	} else {
+		return nil, fmt.Errorf("proxy: should be an object")
+	}
+
+	// Parse the proxy
+	uo := listVal.Filter("upstreams")
+	proxy.Upstreams = make([]*api.ConsulUpstream, len(uo.Items))
+	for i := range uo.Items {
+		u, err := parseUpstream(uo.Items[i])
+		if err != nil {
+			return nil, err
+		}
+
+		proxy.Upstreams[i] = u
+	}
+
+	// If we have config, then parse that
+	if o := listVal.Filter("config"); len(o.Items) > 1 {
+		return nil, fmt.Errorf("only 1 meta object supported")
+	} else if len(o.Items) == 1 {
+		var mSlice []map[string]interface{}
+		if err := hcl.DecodeObject(&mSlice, o.Items[0].Val); err != nil {
+			return nil, err
+		}
+
+		if len(mSlice) > 1 {
+			return nil, fmt.Errorf("only 1 meta object supported")
+		}
+
+		m := mSlice[0]
+
+		if err := mapstructure.WeakDecode(m, &proxy.Config); err != nil {
+			return nil, err
+		}
+
+		proxy.Config = flattenMapSlice(proxy.Config)
+	}
+
+	return &proxy, nil
+}
+
+func parseUpstream(uo *ast.ObjectItem) (*api.ConsulUpstream, error) {
+	valid := []string{
+		"destination_name",
+		"local_bind_port",
+	}
+
+	if err := helper.CheckHCLKeys(uo.Val, valid); err != nil {
+		return nil, multierror.Prefix(err, "upstream ->")
+	}
+
+	var upstream api.ConsulUpstream
+	var m map[string]interface{}
+	if err := hcl.DecodeObject(&m, uo.Val); err != nil {
+		return nil, err
+	}
+
+	dec, err := mapstructure.NewDecoder(&mapstructure.DecoderConfig{
+		DecodeHook:       mapstructure.StringToTimeDurationHookFunc(),
+		WeaklyTypedInput: true,
+		Result:           &upstream,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	if err := dec.Decode(m); err != nil {
+		return nil, err
+	}
+
+	return &upstream, nil
+}
 func parseChecks(service *api.Service, checkObjs *ast.ObjectList) error {
 	service.Checks = make([]api.ServiceCheck, len(checkObjs.Items))
 	for idx, co := range checkObjs.Items {
@@ -1481,162 +1689,6 @@ func parseCheckRestart(cro *ast.ObjectItem) (*api.CheckRestart, error) {
 	}
 
 	return &checkRestart, nil
-}
-
-func parseConnect(co *ast.ObjectItem) (*api.ConsulConnect, error) {
-	valid := []string{
-		"sidecar_service",
-	}
-
-	if err := helper.CheckHCLKeys(co.Val, valid); err != nil {
-		return nil, multierror.Prefix(err, "connect ->")
-	}
-
-	var connect api.ConsulConnect
-
-	var connectList *ast.ObjectList
-	if ot, ok := co.Val.(*ast.ObjectType); ok {
-		connectList = ot.List
-	} else {
-		return nil, fmt.Errorf("connect should be an object")
-	}
-
-	// Parse the sidecar_service
-	o := connectList.Filter("sidecar_service")
-	if len(o.Items) == 0 {
-		return nil, nil
-	}
-	if len(o.Items) > 1 {
-		return nil, fmt.Errorf("only one 'sidecar_service' block allowed per task")
-	}
-
-	r, err := parseSidecarService(o.Items[0])
-	if err != nil {
-		return nil, fmt.Errorf("sidecar_service, %v", err)
-	}
-	connect.SidecarService = r
-
-	return &connect, nil
-}
-
-func parseSidecarService(o *ast.ObjectItem) (*api.ConsulSidecarService, error) {
-	valid := []string{
-		"port",
-		"proxy",
-	}
-
-	if err := helper.CheckHCLKeys(o.Val, valid); err != nil {
-		return nil, multierror.Prefix(err, "sidecar_service ->")
-	}
-
-	var sidecar api.ConsulSidecarService
-	var m map[string]interface{}
-	if err := hcl.DecodeObject(&m, o.Val); err != nil {
-		return nil, err
-	}
-
-	delete(m, "proxy")
-
-	dec, err := mapstructure.NewDecoder(&mapstructure.DecoderConfig{
-		DecodeHook:       mapstructure.StringToTimeDurationHookFunc(),
-		WeaklyTypedInput: true,
-		Result:           &sidecar,
-	})
-	if err != nil {
-		return nil, err
-	}
-	if err := dec.Decode(m); err != nil {
-		return nil, fmt.Errorf("foo: %v", err)
-	}
-
-	var proxyList *ast.ObjectList
-	if ot, ok := o.Val.(*ast.ObjectType); ok {
-		proxyList = ot.List
-	} else {
-		return nil, fmt.Errorf("sidecar_service: should be an object")
-	}
-
-	// Parse the proxy
-	po := proxyList.Filter("proxy")
-	if len(po.Items) == 0 {
-		return &sidecar, nil
-	}
-	if len(po.Items) > 1 {
-		return nil, fmt.Errorf("only one 'proxy' block allowed per task")
-	}
-
-	r, err := parseProxy(po.Items[0])
-	if err != nil {
-		return nil, fmt.Errorf("proxy, %v", err)
-	}
-	sidecar.Proxy = r
-
-	return &sidecar, nil
-}
-
-func parseProxy(o *ast.ObjectItem) (*api.ConsulProxy, error) {
-	valid := []string{
-		"upstreams",
-	}
-
-	if err := helper.CheckHCLKeys(o.Val, valid); err != nil {
-		return nil, multierror.Prefix(err, "proxy ->")
-	}
-
-	var proxy api.ConsulProxy
-
-	var listVal *ast.ObjectList
-	if ot, ok := o.Val.(*ast.ObjectType); ok {
-		listVal = ot.List
-	} else {
-		return nil, fmt.Errorf("proxy: should be an object")
-	}
-
-	// Parse the proxy
-	uo := listVal.Filter("upstreams")
-	proxy.Upstreams = make([]*api.ConsulUpstream, len(uo.Items))
-	for i := range uo.Items {
-		u, err := parseUpstream(uo.Items[i])
-		if err != nil {
-			return nil, err
-		}
-
-		proxy.Upstreams[i] = u
-	}
-
-	return &proxy, nil
-}
-
-func parseUpstream(uo *ast.ObjectItem) (*api.ConsulUpstream, error) {
-	valid := []string{
-		"destination_name",
-		"local_bind_port",
-	}
-
-	if err := helper.CheckHCLKeys(uo.Val, valid); err != nil {
-		return nil, multierror.Prefix(err, "upstream ->")
-	}
-
-	var upstream api.ConsulUpstream
-	var m map[string]interface{}
-	if err := hcl.DecodeObject(&m, uo.Val); err != nil {
-		return nil, err
-	}
-
-	dec, err := mapstructure.NewDecoder(&mapstructure.DecoderConfig{
-		DecodeHook:       mapstructure.StringToTimeDurationHookFunc(),
-		WeaklyTypedInput: true,
-		Result:           &upstream,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	if err := dec.Decode(m); err != nil {
-		return nil, err
-	}
-
-	return &upstream, nil
 }
 
 func parseResources(result *api.Resources, list *ast.ObjectList) error {
