+++ jobspec/test-fixtures/tg-network.hcl
@@ -9,6 +9,25 @@ job "foo" {
         to = 8080
       }
     }
+      
+    service {
+      name        = "connect-service"
+      tags        = ["foo", "bar"]
+      canary_tags = ["canary", "bam"]
+      port        = "1234"
+
+      connect {
+        sidecar_service {
+          proxy {
+            upstreams {
+              destination_name = "other-service"
+              local_bind_port  = 4567
+            }
+          }
+        }
+      }
+    }
+
     task "bar" {
       driver = "raw_exec"
       config {
